+++ secrets/ed25519_test.go
@@ -4,4 +4,5 @@ import "testing"
 
 func TestEd25519(t *testing.T) {
 	// TODO: GenerateEd25519Keys
+	// TODO: LoadEd25519PrivateKey
 }